refactor: :sparkles: añado actualizar_ campo y actualizar_varios
